Add fitness function
The fitness function has been implemented. I ended up taking into
account both correctness and tileset size as it seemed to order the
population better. I also had to remove the autorun as it has a chance
of creating bad data when the interrupt occurs. Another solution will
have to be implemented.